Use ref byte for compression input pointers (#56)
Motivation
----------
Trying to eliminate the use of pinning to avoid interfering with GC.

Modifications
-------------
Replace all input pointers in the compressor with byte ref.

+++ Snappier/Internal/Constants.cs
@@ -20,7 +20,7 @@
 
         public const int BlockLog = 16;
         public const long BlockSize = 1 << BlockLog;
-        public const long InputMarginBytes = 15;
+        public const nint InputMarginBytes = 15;
 
         /// <summary>
         /// Data stored per entry in lookup table:
